% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfhl_functions.R
\name{nfhl_describe}
\alias{nfhl_describe}
\title{Query meta data of NFHL by layer ID}
\usage{
nfhl_describe(layer = 28)
}
\arguments{
\item{layer}{the layer ID to query}
}
\value{
a list containg the layer Name, Description, and bounding box
}
\description{
Query meta data of NFHL by layer ID
}
