% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{nfhl_meta}
\alias{nfhl_meta}
\title{National Flood Hazard Layer Metadata}
\format{
An object of class \code{data.frame} with 32 rows and 2 columns.
}
\usage{
nfhl_meta
}
\description{
A data.frame specifying Layer ID for each date type.
}
\keyword{datasets}
