% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfhl_functions.R
\docType{data}
\name{nhfl_end_point}
\alias{nhfl_end_point}
\title{National Flood Hazard Layer Endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
nhfl_end_point
}
\description{
A character string specifying the base URL for the NFHL web service.
}
\keyword{datasets}
